gateway_service/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
PyJWT